requirements.txt
amqp==5.3.1
anyio==4.5.2
asgiref==3.8.1
attrs==25.3.0
backports.zoneinfo==0.2.1
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
django==4.2.23
django-celery-beat==2.8.1
django-restframework==0.0.1
django-timezone-field==7.1
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-yasg==1.21.10
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib-resources==6.4.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.5.4
packaging==25.0
pkgutil-resolve-name==1.3.10
prompt-toolkit==3.0.51
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.4
rpds-py==0.20.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
zipp==3.20.2
